♻️ refactor : QueryProvider.tsx 이름 수정
