+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
 
 import GlobalModals from "@/app/_components/GlobalModals";
-import QueryProvider from "@/app/_components/QureyProvider";
+import QueryProvider from "@/app/_components/QueryProvider";
 
 export default function Layout({ children }: Readonly<React.PropsWithChildren>) {
 	return (
